+++ examples/print_pos.rs
@@ -10,7 +10,7 @@ fn main() {
     let text = std::fs::read_to_string(&args[1]).unwrap();
     let opt = roxmltree::ParsingOptions {
         allow_dtd: true,
-        nodes_limit: 0,
+        ..roxmltree::ParsingOptions::default()
     };
     let doc = match roxmltree::Document::parse_with_options(&text, opt) {
         Ok(doc) => doc,